Primary screen pipeline
Cell Profiler Pipeline used for analysis of primary screen data. It is important to note that each batch of images needs some optimization to counter for differences in staining. This is done through small changes in the thresholding of the primary objects for infected cells. The cell nucleus staining analysis is usually sufficient as is.